tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/138.html

<!DOCTYPE html>
<html><head>
	<title>scheduler: SVG script nested inlines</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>
  var t = async_test(undefined, {timeout:3000});
</script>
<svg>
<script>
t.step(function() {
log("inline script #2");
var a = {
  <script>
    t.step(function() {log("inline script #1")})
  </script>
a:1}
log("end inline script #2");
});
</script>
</svg>
<script>
onload = t.step_func(function() {
  assert_array_equals(eventOrder, ["inline script #1", "inline script #2", "end inline script #2"]);
  t.done();
});
</script>
